removed all dead code, added light dark theme, largely expanded AI capabilities (from AI overviews, to AI data reads for FBI crime datasets based on zip code or city name (or both for further accuracy)!!!, also bug fixes

+++ lib/models/category_overview_slice.dart
@@ -12,5 +12,5 @@ class CategoryOverviewBarSlice {
   final double share;
   final double riskScore;
   final List<String> factorLabels;
-  final List<int> factorLevels;
+  final List<double> factorLevels;
 }
